Refactored and renamed
