+++ exasol/mlflow_plugin/rest_api/data/__init__.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from exasol.mlflow_plugin.rest_api.column import Column
+from exasol.mlflow_plugin.rest_api.data.column import Column
 
 JsonObject = dict[str, Any]
